イベントルームをリアタイルームのUI基盤に統合し、作品ページにイベントタブを追加
/events/[id] を独自の簡易UIから LiveRoomView（新規共有コンポーネント）経由の
実況ルームと同じ体験（ライブ更新・LiveRoomPostCard・新着/古い順切替）に統合。
放送日時ウィンドウ制限はイベントルームには意図的に適用しない。

作品詳細ページに常時表示の「ルームログ / イベント」タブを追加し、紐づく
イベント一覧を表示できるようにした。

スキーマ側では posts.event_id・events.updated_at・作成者/管理者更新ポリシー・
event_mutes・event_notification_settings、および room_experiment_runs/visits・
room_exit_survey_responses の room_kind 判別列を追加済み（イベント室検証対応は次段階）。
