+++ src/routes/rooms/anime/[id]/[date]/+page.server.ts
@@ -99,7 +99,7 @@ export const load: PageServerLoad = async ({ params, locals: { supabase, safeGet
 		room: {
 			session_id: session.id,
 			date: params.date,
-			kind: "episode",
+			kind: "episode" as const,
 			hashtag,
 			scheduled_at: session.scheduled_at,
 			posting_opens_at: session.posting_opens_at,
