+++ src/routes/rooms/anime/[id]/lobby/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-import RoomPage from "../[date]/+page.svelte";
+import LiveRoomView from "$lib/components/room/LiveRoomView.svelte";
 import type { ActionData, PageData } from "./$types";
 
 let { data, form }: { data: PageData; form: ActionData } = $props();
 </script>
 
-<RoomPage {data} {form} />
+<LiveRoomView {data} {form} />
